+++ reactivestate-core/src/commonMain/kotlin/com/ensody/reactivestate/OnDemandStateFlow.kt
@@ -1,12 +1,16 @@
 package com.ensody.reactivestate
 
-import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.NonCancellable
+import kotlinx.coroutines.channels.ProducerScope
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.FlowCollector
-import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.flow.shareIn
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
@@ -16,80 +20,78 @@ import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
 
 /**
- * Creates a computed [StateFlow] without requiring a [CoroutineScope] (unlike [stateIn]).
+ * Turns this [Flow] into a [StateFlow] without requiring a [CoroutineScope] (unlike [stateIn]).
  *
- * This can be useful for callback based observers where you only want to register the [observer] when this [StateFlow]
- * gets collected. When nobody collects, the [StateFlow.value] gets computed via [getter].
+ * When the resulting [StateFlow] gets collected this in turn starts collecting this underlying [Flow].
+ * When nobody collects, the [StateFlow.value] gets computed via [getter] which is also passed the previous value
+ * wrapped in a [Wrapped] instance or null if there is no previous value, yet.
+ * This way you can implement simple caching.
  *
- * The [observer] receives an [OnDemandStateFlowContext] as `this` and should keep the [OnDemandStateFlowContext.value]
- * constantly updated. You can either use `try`-`finally` or [OnDemandStateFlowContext.invokeOnCancellation] to clean up
- * when all collectors have stopped.
+ * When nobody collects, this is safe for garbage collection.
+ *
+ * Normally you'd use this together with [callbackFlow], for example.
  */
-public fun <T> OnDemandStateFlow(
-    getter: MutableStateFlow<T>?.() -> T,
+public fun <T> Flow<T>.stateOnDemand(
     context: CoroutineContext = EmptyCoroutineContext,
-    observer: suspend OnDemandStateFlowContext<T>.() -> Unit,
+    getter: (previous: Wrapped<T>?) -> T,
 ): StateFlow<T> =
-    DefaultOnDemandStateFlow(
-        getter = getter,
-        observer = observer,
-        delegate = MutableStateFlow(null.getter()),
-        context = context,
-    )
+    DefaultOnDemandStateFlow(internalShareOnDemand(replay = 1, context = context), getter)
 
 /**
- * Passed to [OnDemandStateFlow]'s `observer` which must keep [value] constantly updated.
+ * Turns this [Flow] into a [SharedFlow] without requiring a [CoroutineScope] (unlike [shareIn]).
  *
- * You can either use `try`-`finally` or [OnDemandStateFlowContext.invokeOnCancellation] to clean up when
- * all collectors have stopped.
+ * The underlying [Flow] is only collected while there is at least one collector.
+ * When nobody collects, this is safe for garbage collection.
+ *
+ * Normally you'd use this together with [callbackFlow], for example.
+ * Alternatively you can construct a [SharedFlow] directly via [sharedFlow].
  */
-public class OnDemandStateFlowContext<T>(private val delegate: MutableStateFlow<T>) {
-    public var value: T
-        get() = delegate.value
-        set(value) { delegate.value = value }
+public fun <T> Flow<T>.shareOnDemand(
+    replay: Int = 0,
+    context: CoroutineContext = EmptyCoroutineContext,
+): SharedFlow<T> =
+    internalShareOnDemand(replay = replay, context = context)
 
-    internal var cancellationBlock: (suspend () -> Unit)? = null
+/**
+ * Creates a computed [SharedFlow] without requiring a [CoroutineScope] (unlike [shareIn]).
+ *
+ * The [observer] block is only executed while there is at least one collector.
+ * When nobody collects, this is safe for garbage collection.
+ */
+public inline fun <T> sharedFlow(
+    replay: Int = 0,
+    context: CoroutineContext = EmptyCoroutineContext,
+    crossinline observer: suspend ProducerScope<T>.() -> Unit,
+): SharedFlow<T> =
+    callbackFlow { observer() }.shareOnDemand(replay = replay, context = context)
 
-    public suspend fun awaitCancellation() {
-        CompletableDeferred<Unit>().await()
+private fun <T> Flow<T>.internalShareOnDemand(
+    replay: Int = 0,
+    context: CoroutineContext = EmptyCoroutineContext,
+): DefaultOnDemandSharedFlow<T> =
+    if (this is DefaultOnDemandSharedFlow<T> && this.replay == replay && this.context == context) {
+        this
+    } else {
+        DefaultOnDemandSharedFlow(this, replay, context, MutableSharedFlow(replay = replay))
     }
 
-    public fun invokeOnCancellation(block: suspend () -> Unit) {
-        cancellationBlock = block
-    }
-}
-
-private class DefaultOnDemandStateFlow<T>(
-    private val getter: MutableStateFlow<T>?.() -> T,
-    private val observer: suspend OnDemandStateFlowContext<T>.() -> Unit,
-    private val delegate: MutableStateFlow<T>,
-    private val context: CoroutineContext = EmptyCoroutineContext,
-) : StateFlow<T> by delegate {
+private class DefaultOnDemandSharedFlow<T>(
+    val flow: Flow<T>,
+    val replay: Int,
+    val context: CoroutineContext,
+    val delegate: MutableSharedFlow<T>,
+) : SharedFlow<T> by delegate {
 
     private val mutex = Mutex()
     private var job: Job? = null
-    private var subscriptionCount = 0
-
-    override val value: T get() = if (subscriptionCount == 0) delegate.getter() else delegate.value
-
-    private fun flowContext() = OnDemandStateFlowContext(delegate)
+    var subscriptionCount = 0
 
     override suspend fun collect(collector: FlowCollector<T>): Nothing {
-        if (delegate.subscriptionCount.value == 0) {
-            delegate.value = delegate.getter()
-        }
         mutex.withLock {
             subscriptionCount += 1
             if (job == null) {
                 job = mainScope.launch(context = context) {
-                    val flowContext = flowContext()
-                    try {
-                        flowContext.observer()
-                    } finally {
-                        withContext(NonCancellable) {
-                            flowContext.cancellationBlock?.invoke()
-                        }
-                    }
+                    flow.collect { delegate.emit(it) }
                 }
             }
         }
@@ -106,5 +108,32 @@ private class DefaultOnDemandStateFlow<T>(
                 }
             }
         }
+    }
+}
+
+private class DefaultOnDemandStateFlow<T>(
+    private val delegate: DefaultOnDemandSharedFlow<T>,
+    private val getter: (previous: Wrapped<T>?) -> T,
+) : StateFlow<T>, SharedFlow<T> by delegate {
+
+    override val value: T
+        get() =
+            if (delegate.subscriptionCount == 0 || delegate.replayCache.isEmpty()) {
+                val previous = if (delegate.replayCache.isNotEmpty()) Wrapped(delegate.replayCache.first()) else null
+                getter(previous).also {
+                    if (previous == null || it !== previous.value) {
+                        delegate.delegate.tryEmit(it)
+                    }
+                }
+            } else {
+                delegate.replayCache.first()
+            }
+
+    override suspend fun collect(collector: FlowCollector<T>): Nothing {
+        // Ensure the first emitted value is up to date
+        if (delegate.subscriptionCount == 0) {
+            value
+        }
+        delegate.collect(collector)
     }
 }